add features

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\MoviesController;
+use App\Http\Controllers\MoviesUsersController;
 use App\Http\Controllers\UsersController;
 use Illuminate\Support\Facades\Route;
 
@@ -15,30 +16,48 @@ use Illuminate\Support\Facades\Route;
 |
  */
 
+//Get Requests
+
 Route::get('/', [MoviesController::class, 'index']);
 
-Route::get('/movies/create', [MoviesController::class, 'create']);
+Route::get('/movies', function () {
+    return redirect('/');
+});
 
-Route::post('/movies', [MoviesController::class, 'store']);
+Route::get('/movies/create', [MoviesController::class, 'create'])->middleware('auth');
 
 Route::get('/movies/{movie}', [MoviesController::class, 'show']);
 
-Route::get('/movies/{movie}/edit', [MoviesController::class, 'edit']);
+Route::get('/movies/{movie}/edit', [MoviesController::class, 'edit'])->middleware('auth');
 
-Route::delete('/movies/{movie}', [MoviesController::class, 'destroy']);
+Route::get('/register', [UsersController::class, 'create'])->middleware('guest');
 
-Route::put('/movies/{movie}', [MoviesController::class, 'update']);
+Route::get('/users/{user}', [UsersController::class, 'show'])->middleware('auth');
 
-Route::get('/register', [UsersController::class, 'create']);
+Route::get('/users/{user}/edit', [UsersController::class, 'edit'])->middleware('auth');
 
-Route::get('/users', [UsersController::class, 'index']);
+Route::get('/login', [UsersController::class, 'login'])->name('login')->middleware('guest');
 
-Route::post('/users', [UsersController::class, 'store']);
+Route::get('/users', [UsersController::class, 'index'])->middleware('auth');
 
-Route::get('/users/{user}', [UsersController::class, 'show']);
+//Post Requests
 
-Route::get('/users/{user}/edit', [UsersController::class, 'edit']);
+Route::post('/movies', [MoviesController::class, 'store'])->middleware('auth');
 
-Route::delete('/users/{user}', [UsersController::class, 'destroy']);
+Route::delete('/movies/{movie}', [MoviesController::class, 'destroy'])->middleware('auth');
 
-Route::put('/users/{user}', [UsersController::class, 'update']);
+Route::put('/movies/{movie}', [MoviesController::class, 'update'])->middleware('auth');
+
+Route::post('/users', [UsersController::class, 'store'])->middleware('guest');
+
+Route::delete('/users/{user}', [UsersController::class, 'destroy'])->middleware('auth');
+
+Route::put('/users/{user}', [UsersController::class, 'update'])->middleware('auth');
+
+Route::post('/logout', [UsersController::class, 'logout'])->middleware('auth');
+
+Route::post('/users/authenticate', [UsersController::class, 'authenticate'])->middleware('guest');
+
+Route::post('/like', [MoviesUsersController::class, 'store']);
+
+Route::delete('/dislike', [MoviesUsersController::class, 'destroy']);
